+++ scriptassistant/tests/test_script_assistant.py
@@ -1,13 +1,10 @@
 # -*- coding: utf-8 -*-
 
 import os
-import sys
-import time
 import unittest
 
-from qgis.core import QgsApplication, QgsMapLayerRegistry
-from qgis.gui import QgsMessageBar
-from qgis.utils import plugins, iface
+from qgis.core import QgsApplication
+from qgis.utils import plugins
 
 __location__ = os.path.realpath(
     os.path.join(os.getcwd(), os.path.dirname(__file__)))
@@ -34,20 +31,11 @@ class ScriptAssistantSettingsTest(unittest.TestCase):
         self.dlg.show()
         self.dlg.populate_config_combo()
         self.dlg.show_last_configuration()
-        self.view(self.dlg)
 
     def tearDown(self):
         """Runs after each test."""
         self.dlg.close()
-        self.view(self.dlg)
         pass
-
-    def view(self, qt_object):
-        if view_tests_during_execution:
-            qt_object.repaint()
-            time.sleep(1)
-        else:
-            pass
 
     def test_correct_script_assistant_test_settings(self):
         self.assertEqual(self.dlg.cmb_config.lineEdit().text(), "Script Assistant")
@@ -61,7 +49,6 @@ class ScriptAssistantSettingsTest(unittest.TestCase):
     def test_deleting_settings(self):
         count = self.dlg.cmb_config.count()
         self.dlg.btn_delete.clicked.emit(True)
-        self.view(self.dlg)
         if count == 1:
             self.assertTrue(self.dlg.btn_save.isEnabled())
             self.assertFalse(self.dlg.btn_delete.isEnabled())
@@ -80,15 +67,12 @@ class ScriptAssistantSettingsTest(unittest.TestCase):
         self.dlg.chk_reload.setChecked(False)
         self.dlg.chk_repaint.setChecked(True)
         self.dlg.btn_save.clicked.emit(True)
-        self.view(self.dlg)
 
     def test_adding_new_settings(self):
         self.dlg.cmb_config.lineEdit().setText("Script Assistant Test")
-        self.view(self.dlg)
         self.assertIn("*", self.dlg.windowTitle())
         self.dlg.lne_test_data.setText(os.path.join(__location__, "tests", "testdata"))
         self.assertTrue(self.dlg.btn_save.isEnabled())
-        self.view(self.dlg)
         self.dlg.btn_save.clicked.emit(True)
         self.assertEqual(self.dlg.cmb_config.lineEdit().text(), "Script Assistant Test")
         self.assertEqual(self.dlg.lne_test_data.text(), os.path.join(__location__, "tests", "testdata"))
@@ -96,24 +80,20 @@ class ScriptAssistantSettingsTest(unittest.TestCase):
         self.assertTrue(self.dlg.btn_delete.isEnabled())
         self.dlg.btn_delete.clicked.emit(True)
         self.dlg.cmb_config.setCurrentIndex(self.dlg.cmb_config.findText("Script Assistant"))
-        self.view(self.dlg)
 
     def test_running_script_test(self):
-        self.view(self.dlg)
         self.dlg.lne_test.setText(os.path.join(__location__, "tests"))
         self.assertIn("*", self.dlg.windowTitle())
         self.dlg.btn_save.clicked.emit(True)
         self.assertNotIn("*", self.dlg.windowTitle())
         self.scriptassistant.save_settings()
         self.dlg.close()
-        self.view(self.dlg)
         for action in self.scriptassistant.test_actions:
             if action.text() == "add_area_column":
                 action.triggered.emit(True)
         self.dlg.show()
         self.dlg.populate_config_combo()
         self.dlg.show_last_configuration()
-        self.view(self.dlg)
         self.dlg.lne_test.setText(__location__)
         self.dlg.lne_test_data.setText("")
         self.dlg.btn_save.clicked.emit(True)
@@ -127,45 +107,19 @@ class ScriptAssistantSettingsTest(unittest.TestCase):
         plugins["processing"].toolbox.updateProvider("script")
 
     def test_running_plugin_test(self):
-        self.view(self.dlg)
         self.dlg.lne_test.setText(os.path.join(__location__, "tests"))
         self.assertIn("*", self.dlg.windowTitle())
         self.dlg.btn_save.clicked.emit(True)
         self.assertNotIn("*", self.dlg.windowTitle())
         self.scriptassistant.save_settings()
         self.dlg.close()
-        self.view(self.dlg)
         for action in self.scriptassistant.test_actions:
             if action.text() == "plugin":
                 action.triggered.emit(True)
         self.dlg.show()
         self.dlg.populate_config_combo()
         self.dlg.show_last_configuration()
-        self.view(self.dlg)
         self.dlg.lne_test.setText(__location__)
         self.dlg.lne_test_data.setText("")
         self.dlg.btn_save.clicked.emit(True)
         self.scriptassistant.save_settings()
-
-
-def run_tests(view_tests=False):
-    global view_tests_during_execution
-    if view_tests:
-        view_tests_during_execution = True
-    else:
-        view_tests_during_execution = False
-    suite = unittest.TestSuite()
-    suite.addTests(unittest.makeSuite(ScriptAssistantSettingsTest, "test"))
-    result = unittest.TextTestRunner(verbosity=2, stream=sys.stdout).run(suite)
-    if result.wasSuccessful():
-        iface.messageBar().pushMessage(
-            "All Tests Passed",
-            "Testing was successful.",
-            level=QgsMessageBar.SUCCESS,
-        )
-    else:
-        iface.messageBar().pushMessage(
-            "Test Failure",
-            "Testing was not successful.",
-            level=QgsMessageBar.CRITICAL,
-        )